Refactor aside menu to styled components
