+++ src/components-styled/aside/title-with-icon.tsx
