+++ src/components/layout/index.tsx
@@ -76,11 +76,12 @@ function Layout(
         url={url}
       />
 
-      <div className={styles.skiplinks}>
+      <nav role="navigation" className={styles.skiplinks}>
         <a href="#content">{text.skiplinks.inhoud}</a>
         <a href="#main-navigation">{text.skiplinks.nav}</a>
+        <a href="#metric-navigation">{text.skiplinks.metricnav}</a>
         <a href="#footer-navigation">{text.skiplinks.footernav}</a>
-      </div>
+      </nav>
 
       <header className={styles.header}>
         <div className={styles.logoWrapper}>
