+++ src/locale/nl.json
@@ -28,6 +28,7 @@
   "skiplinks": {
     "inhoud": "Ga direct naar de inhoud",
     "nav": "Ga direct naar de navigatie",
+    "metricnav": "Ga direct naar de metrieken navigatie",
     "footernav": "Ga direct naar de footer navigatie"
   },
   "header": {
