+++ src/scss/layout.scss
@@ -333,28 +333,6 @@
   background: white;
   box-shadow: 1px -1px 5px #e5e5e5, -1px -1px 5px #e5e5e5;
   z-index: 2;
-
-  h2 {
-    font-size: 1.42383rem;
-    line-height: 1.40466;
-    padding: 1rem;
-    padding-bottom: 0.5rem;
-    padding-top: 2rem;
-  }
-
-  ul {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-
-    li {
-      border-bottom: 1px solid #c4c4c4;
-
-      &:first-child {
-        border-top: 1px solid #c4c4c4;
-      }
-    }
-  }
 }
 
 .metric-wrapper {
@@ -364,18 +342,6 @@
   & > div:not(:first-child) {
     height: 3.5rem;
     margin-top: -1.25em;
-  }
-}
-
-.region-names {
-  h2 {
-    padding-bottom: 0;
-    margin-bottom: 0;
-  }
-
-  p {
-    margin-top: 0;
-    padding: 0 1rem 1rem;
   }
 }
